refactor: relocate platform type definition and update imports
