+++ src/components/PlatformSelector.tsx
@@ -1,7 +1,6 @@
 import { Button, Menu, MenuButton, MenuItem, MenuList } from '@chakra-ui/react'
 import { BsChevronDown } from 'react-icons/bs'
-import usePlatforms from '../hooks/usePlatforms'
-import { Platform } from '../hooks/useGames'
+import usePlatforms, { type Platform } from '../hooks/usePlatforms'
 
 type Props = {
   onSelectPlatform: (platform: Platform) => void
